refactor: View와 Rest Controller 분리
