+++ src/main/java/cart/catalog/view/CatalogController.java
@@ -1,4 +1,4 @@
-package cart.catalog.controller;
+package cart.catalog.view;
 
 import cart.catalog.service.CatalogService;
 import org.springframework.stereotype.Controller;
@@ -7,18 +7,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class CatalogController {
-    
+
     private final CatalogService catalogService;
-    
+
     public CatalogController(final CatalogService catalogService) {
         this.catalogService = catalogService;
     }
-    
+
     @GetMapping("/")
     public String renderProductListPage(final Model model) {
         model.addAttribute("products", this.catalogService.display());
         return "index";
     }
-    
-    
 }
