+++ src/main/java/cart/catalog/view/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package cart.catalog.controller;
+package cart.catalog.view;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 @ControllerAdvice
 public class ExceptionAdvice {
-    
+
     @ExceptionHandler({Exception.class})
     public ResponseEntity<String> handleAll(final Exception exception) {
         return ResponseEntity.badRequest().body(exception.getMessage());
